% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{add_annotation_redeem}
\alias{add_annotation_redeem}
\title{add_annotation_redeem}
\usage{
add_annotation_redeem(ob)
}
\arguments{
\item{ob}{redeem object}

\item{fdr}{fdr cutoff, default is 0.05}
}
\description{
This function is to simply add all annotation in
}
