% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{Add_AssignVariant,mitoTracing-method}
\alias{Add_AssignVariant,mitoTracing-method}
\title{a function to assign variants to edges based on maximum likihood}
\usage{
\S4method{Add_AssignVariant}{mitoTracing}(mitoTracing = DN1_HSC_mitoTracing.VerySensitive, n.cores = 4)
}
\arguments{
\item{mitoTracing}{Need to have mitoTracing@Ctx.Mtx.depth (By Add_DepthMatrix),  mitoTracing@Cts.Mtx  mitoTracing@Cts.Mtx.bi, mitoTracing@TREE}
}
\value{
mitoTracing with @AssignedVarian list of two p is a probability matrix of variants vs edges (Rowsum is 1) and Variant.assign.report, a dataframe (Variant|Edge.Assign|prob)
}
\description{
a function to assign variants to edges based on maximum likihood
}
