% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{Add_AssignVariant,redeemR-method}
\alias{Add_AssignVariant,redeemR-method}
\title{a function to assign variants to edges based on maximum likihood}
\usage{
\S4method{Add_AssignVariant}{redeemR}(redeemR = DN1_HSC_redeemR.VerySensitive, n.cores = 4)
}
\arguments{
\item{redeemR}{Need to have redeemR@Ctx.Mtx.depth (By Add_DepthMatrix),  redeemR@Cts.Mtx  redeemR@Cts.Mtx.bi, redeemR@TREE}
}
\value{
redeemR with @AssignedVarian list of two p is a probability matrix of variants vs edges (Rowsum is 1) and Variant.assign.report, a dataframe (Variant|Edge.Assign|prob)
}
\description{
a function to assign variants to edges based on maximum likihood
}
