% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{Add_AssignVariant}
\alias{Add_AssignVariant}
\title{Add_AssignVariant
a function to assign variants to edges based on maximum likihood}
\usage{
Add_AssignVariant(redeemR, n.cores, ...)
}
\arguments{
\item{object}{redeemR class}

\item{QualifiedTotalCts}{a big source data, usually at XXX/mitoV/final}
}
\description{
Add_AssignVariant
a function to assign variants to edges based on maximum likihood
}
