% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{add_changes}
\alias{add_changes}
\title{Extract nucleotide change from variant string}
\usage{
add_changes(variant)
}
\arguments{
\item{variant}{Character vector of variant strings ("\if{html}{\out{<pos>}}\emph{\if{html}{\out{<ref>}}}\if{html}{\out{<alt>}}").}
}
\value{
Character vector of nucleotide changes ("\if{html}{\out{<ref>}}_\if{html}{\out{<alt>}}").
}
\description{
Given a variant in the format "\if{html}{\out{<pos>}}\emph{\if{html}{\out{<ref>}}}\if{html}{\out{<alt>}}", returns the substring representing
the reference-to-alternate change (e.g. "A_T").
}
\examples{
add_changes("150_A_G")  # returns "A_G"
}
