% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{Add_DepthMatrix,redeemR-method}
\alias{Add_DepthMatrix,redeemR-method}
\title{Add_DepthMatrix
Optional, add a matrix with same dimension with the Cts.Mtx and Cts.Mtx.bi, which display the depths}
\usage{
\S4method{Add_DepthMatrix}{redeemR}(object, QualifiedTotalCts = "")
}
\arguments{
\item{object}{redeemR class}

\item{QualifiedTotalCts}{a big source data, usually at XXX/mitoV/final,  If needed, edit V1, the cell name, which may have additional postfix due to combine}
}
\value{
redeemR class
}
\description{
Add_DepthMatrix
Optional, add a matrix with same dimension with the Cts.Mtx and Cts.Mtx.bi, which display the depths
}
