% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{Add_DepthMatrix}
\alias{Add_DepthMatrix}
\title{Add_DepthMatrix
Optional, add a matrix with same dimension with the Cts.Mtx and Cts.Mtx.bi, which display the depths}
\usage{
Add_DepthMatrix(object, QualifiedTotalCts, ...)
}
\arguments{
\item{object}{mitoTracin class}

\item{QualifiedTotalCts}{a big source data, usually at XXX/mitoV/final}
}
\description{
Add_DepthMatrix
Optional, add a matrix with same dimension with the Cts.Mtx and Cts.Mtx.bi, which display the depths
}
