% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{add_derived_profile_info}
\alias{add_derived_profile_info}
\title{This is a convinience function, internal borrowed from https://github.com/NickWilliamsSanger/treemut/blob/main/R/treemut.R#L68}
\usage{
add_derived_profile_info(
  profile_df,
  samples = sprintf("s\%s", 0:(nchar(profile_df$profile[1]) - 1))
)
}
\description{
This is a convinience function, internal borrowed from https://github.com/NickWilliamsSanger/treemut/blob/main/R/treemut.R#L68
}
