% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{add_median_depth_to_redeemR}
\alias{add_median_depth_to_redeemR}
\title{Add median depth information to redeemR object}
\usage{
add_median_depth_to_redeemR(redeemR)
}
\arguments{
\item{redeemR}{A redeemR object that contains a depth matrix in @Ctx.Mtx.depth}
}
\value{
The input redeemR object with median_depth column added to @V.fitered
}
\description{
Computes the median depth for each variant from the depth matrix and adds it
to the V.fitered slot by joining with the variant information.
}
\details{
This function:
\enumerate{
\item Extracts the depth matrix from the redeemR object
\item Computes median depth for each variant (column) in the depth matrix
\item Converts variant names using convert_variant() for proper joining
\item Joins the median depth information with V.fitered
\item Returns the modified redeemR object
}
}
\examples{
redeemR <- add_median_depth_to_redeemR(redeemR)

}
