% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSummary.R
\name{add_raw_fragment}
\alias{add_raw_fragment}
\title{Produce a raw fragment table with frequency (how many cells) and the reletive distance
from redeemR object,}
\usage{
add_raw_fragment(redeemR, raw = "RawGenotypes.Sensitive.StrandBalance")
}
\arguments{
\item{redeemR}{a redeemR object}
}
\description{
Produce a raw fragment table with frequency (how many cells) and the reletive distance
from redeemR object,
}
