% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{Add_tree_cut,redeemR-method}
\alias{Add_tree_cut,redeemR-method}
\title{a function to cut tree using assigned variant as branch-length on edge}
\usage{
\S4method{Add_tree_cut}{redeemR}(
  redeemR = DN4_stemcell_redeemR.seed.verysensitive,
  MinCell = 30,
  N = 1,
  prob.cut = 0.3,
  Dumpcut = 100
)
}
\arguments{
\item{redeemR}{Need to have had the tree built}

\item{MinCell}{The minimum number of cells in each clone, otherwise merge with sibling}

\item{N}{branch length to cut the tree}

\item{Dumpcut}{Number of can be tolerated to be removed to fulfill the right side. The small value-> Less unassignment, big clones}
}
\description{
a function to cut tree using assigned variant as branch-length on edge
}
