% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{AddDatatoplot_clustering,mitoTracing-method}
\alias{AddDatatoplot_clustering,mitoTracing-method}
\title{AddDatatoplot_clustering
This prepare the clonal clustering data to plot}
\usage{
\S4method{AddDatatoplot_clustering}{mitoTracing}(object)
}
\arguments{
\item{object}{mitoTracin class}
}
\value{
mitoTracing class
}
\description{
AddDatatoplot_clustering
This prepare the clonal clustering data to plot
}
