% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{AddDatatoplot_clustering}
\alias{AddDatatoplot_clustering}
\title{AddDatatoplot_clustering
This prepare the clonal clustering data to plot}
\usage{
AddDatatoplot_clustering(object, ...)
}
\arguments{
\item{object}{redeemR class}
}
\description{
AddDatatoplot_clustering
This prepare the clonal clustering data to plot
}
