% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{AddDist,mitoTracing-method}
\alias{AddDist,mitoTracing-method}
\title{AddDist
This add Jaccard, Dice, Jaccard3W distance and stored in DistObjects}
\usage{
\S4method{AddDist}{mitoTracing}(object)
}
\arguments{
\item{object}{mitoTracin class}
}
\value{
mitoTracing class
}
\description{
AddDist
This add Jaccard, Dice, Jaccard3W distance and stored in DistObjects
}
