% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{AddDist}
\alias{AddDist}
\title{AddDist
This add Jaccard, Dice, Jaccard3W distance and stored in DistObjects}
\usage{
AddDist(object, ...)
}
\arguments{
\item{object}{mitoTracin class}
}
\description{
AddDist
This add Jaccard, Dice, Jaccard3W distance and stored in DistObjects
}
