% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{AddDist_legacy}
\alias{AddDist_legacy}
\title{AddDist_legacy
This add Jaccard, Dice, Jaccard3W distance and stored in DistObjects}
\usage{
AddDist_legacy(object, ...)
}
\arguments{
\item{object}{mitoTracing class}
}
\description{
AddDist_legacy
This add Jaccard, Dice, Jaccard3W distance and stored in DistObjects
}
