% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AddHemSignature}
\alias{AddHemSignature}
\title{Function to add hematopoietic signatures from Griffin_Signatures}
\usage{
AddHemSignature(object = Donor01_BMMC_Multiome_wrapper.filtered)
}
\arguments{
\item{object}{a seurat object}
}
\value{
a seurat object
}
\description{
This function allows you to input a seurat object, add the signatures and return an seurat object
}
