% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{AddTree,mitoTracing-method}
\alias{AddTree,mitoTracing-method}
\title{Add_Tree
Optional, if a phylogentic tree object phylo is already available, can be directly added to the mitoTracing class in slot TREE}
\usage{
\S4method{AddTree}{mitoTracing}(object, phylo, record = "")
}
\arguments{
\item{object}{mitoTracin class}

\item{phylo}{phyogenetic tree object}
}
\value{
mitoTracing class
}
\description{
Add_Tree
Optional, if a phylogentic tree object phylo is already available, can be directly added to the mitoTracing class in slot TREE
}
