% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{AddTree,redeemR-method}
\alias{AddTree,redeemR-method}
\title{Add_Tree
Optional, if a phylogentic tree object phylo is already available, can be directly added to the redeemR class in slot TREE}
\usage{
\S4method{AddTree}{redeemR}(object, phylo, record = "")
}
\arguments{
\item{object}{redeemR class}

\item{phylo}{phyogenetic tree object}
}
\value{
redeemR class
}
\description{
Add_Tree
Optional, if a phylogentic tree object phylo is already available, can be directly added to the redeemR class in slot TREE
}
