% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{AddTree}
\alias{AddTree}
\title{Add_Tree
Optional, if a phylogentic tree object phylo is already available, can be directly added to the redeemR}
\usage{
AddTree(object, phylo, ...)
}
\arguments{
\item{object}{redeemR class}

\item{phylo}{phyogenetic tree object}
}
\description{
Add_Tree
Optional, if a phylogentic tree object phylo is already available, can be directly added to the redeemR
}
