% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{annotate_variants_blacklist}
\alias{annotate_variants_blacklist}
\title{Annotate variants with mitochondrial blacklist region flag}
\usage{
annotate_variants_blacklist(variant_annotation)
}
\arguments{
\item{variant_annotation}{A data.frame or tibble containing at least a column
named \code{Variants}, where each entry is of the form \code{"<pos>_<ref>_<alt>"}.}
}
\value{
The input \code{variant_annotation} with an added column:
\describe{
\item{\code{blacklist_region}}{Character; \code{"blacklist_region"} if the variant
position is in one of the predefined blacklist intervals, otherwise \code{""}.}
}
}
\description{
Flags variants whose genomic position falls within known problematic
regions of the mitochondrial genome (blacklist regions).
}
\details{
The blacklist intervals correspond to regions in the revised Cambridge Reference
Sequence (rCRS) prone to misalignment:
\enumerate{
\item Misalignment due to \code{ACCCCCC T C C C C C C} (rCRS 302–315)
\item Misalignment due to \code{G C A C A C A C A C A C C} (rCRS 513–525)
\item Misalignment due to \code{3107N} ambiguity in rCRS (rCRS 3105–3109)
\item Misalignment due to \code{ACCCCC} homopolymer (rCRS 16182–16187)
}
}
