% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_functions.R
\name{ATAC_Wrapper}
\alias{ATAC_Wrapper}
\title{Wrap Seurat ATAC clustering}
\usage{
ATAC_Wrapper(MTX, res = 0.3, dim1 = 1, dim2 = 20)
}
\arguments{
\item{MTX}{sparse Matrix of class "dgCMatrix", each row is a peak, each column is a cell,}

\item{res}{clustering resolution, default=0.5}
}
\value{
this returns seurat object with ATAC clustering
}
\description{
This function allows you to perform standard sc-ATAC clustering
}
\examples{
bmmc.filtered.atac<-SeuratLSIClustering(PeakVSCell.filtered.Mtx) #each row is a peak, each column is a cell,
}
