% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{BinaryDist}
\alias{BinaryDist}
\title{Compute distances for binary distances}
\usage{
BinaryDist(M, method = "Jaccard")
}
\arguments{
\item{M}{the binary matrix, Each row is a cell, each column is a variant, generated by Make_matrix}

\item{method}{distance method, choose from Jaccard, Dice, 3WJaccard, Simpson, Kulczynski2, Ochiai, Hamming}
}
\value{
dist object
}
\description{
Compute distances for binary distances
}
\examples{
m <- Matrix::Matrix(c(1, 0, 1, 1, 0, 1), nrow = 3, sparse = TRUE)
d.Jaccard <- BinaryDist(m, method = "Jaccard")
}
