% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions.R
\name{binomialtest.msig.enrch_deplet}
\alias{binomialtest.msig.enrch_deplet}
\title{binomialtest.msig.enrch_deplet}
\usage{
binomialtest.msig.enrch_deplet(mylist, All = All.genes, name, thedatabase = db)
}
\description{
This function is an internal function calculating the significance
}
