% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{Build_Manhattan_Tree}
\alias{Build_Manhattan_Tree}
\title{Build Manhattan Tree}
\usage{
Build_Manhattan_Tree(heteroplasmy_matrix)
}
\arguments{
\item{matrices}{A list of matrices containing heteroplasmy values}

\item{nj}{Logical indicating whether to build NJ tree (default: TRUE)}
}
\value{
A list of rooted phylo objects
}
\description{
Builds a Manhattan tree from a heteroplasmy matrix
}
