% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{clean_redeem}
\alias{clean_redeem}
\title{clean_redeem}
\usage{
clean_redeem(ob, fdr = 0.05)
}
\arguments{
\item{ob}{redeem object}

\item{fdr}{fdr cutoff, default is 0.05}
}
\description{
This function is to clean redeem by filtering both V.fitered and GTsummary.filtered by qvalues
}
