% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{clean_redeem_remove_blacklist_RSRS50}
\alias{clean_redeem_remove_blacklist_RSRS50}
\title{clean_redeem_remove_blacklist_RSRS50}
\usage{
clean_redeem_remove_blacklist_RSRS50(ob)
}
\arguments{
\item{ob}{redeem object}
}
\description{
This function is to clean redeem by filtering both V.fitered and GTsummary.filtered by qvalues
}
