% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{clean_redeem_removehomo}
\alias{clean_redeem_removehomo}
\title{clean_redeem_removehomo}
\usage{
clean_redeem_removehomo(ob)
}
\arguments{
\item{ob}{redeem object}

\item{hotcall}{fdr cutoff, default is 0.05}
}
\description{
This function is to clean redeem by filtering both V.fitered and GTsummary.filtered by qvalues
}
