% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{clean_redeem_removehot}
\alias{clean_redeem_removehot}
\title{clean_redeem_removehotcall}
\usage{
clean_redeem_removehot(ob, hotcall = c("310_T_C", "9979_G_A", "3109_T_C"))
}
\arguments{
\item{ob}{redeem object}

\item{hotcall}{fdr cutoff, default is 0.05}
}
\description{
This function is to clean redeem by filtering both V.fitered and GTsummary.filtered by qvalues
}
