% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{compute_distance_matrix}
\alias{compute_distance_matrix}
\title{Compute Distance Matrix from Heteroplasmy Matrix}
\usage{
compute_distance_matrix(heteroplasmy_matrix, dist_method = "manhattan")
}
\arguments{
\item{heteroplasmy_matrix}{A matrix of cell-by-variant heteroplasmy values}

\item{dist_method}{The distance method to use: "manhattan" or "euclidean" (default: "manhattan")}
}
\value{
A distance matrix (dist object)
}
\description{
Computes distance matrix between cells using either Manhattan or Euclidean distance.
}
