% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{convert_mitotracing_redeemR}
\alias{convert_mitotracing_redeemR}
\title{convert_mitotracing_redeemR}
\usage{
convert_mitotracing_redeemR(mitotracing, addAssignedVariant = T)
}
\arguments{
\item{mitotracing}{mitotracing object}
}
\value{
redeemR class
}
\description{
This function is to convert the older version mitotracing into the redeemR
}
