% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_format.R
\name{convert_variant}
\alias{convert_variant}
\title{convert_variant}
\usage{
convert_variant(x)
}
\arguments{
\item{x}{this is a vector of variant names, either way for example from '10000_G_A' to/from 'Variants10000GA'}
}
\value{
a vector of strings
}
\description{
convert_variant
}
