% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CountOverlap_Adj}
\alias{CountOverlap_Adj}
\title{CountOverlap_Adj
function to count the connectedness (adjacency matrix), or the number of cells sharing more than n variants with the given cell}
\usage{
CountOverlap_Adj(M, n = 0)
}
\arguments{
\item{M}{}

\item{n}{}
}
\description{
CountOverlap_Adj
function to count the connectedness (adjacency matrix), or the number of cells sharing more than n variants with the given cell
}
