% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{Create_mitoTracing}
\alias{Create_mitoTracing}
\title{Create_mitoTracing}
\usage{
Create_mitoTracing(
  GTsummary_list,
  depth_list,
  feature.list_list,
  meta_list,
  labels,
  thr = "VerySensitive",
  qualifiedCellCut = 10,
  OnlyHetero = T,
  VAFcut = 1,
  Cellcut = 2,
  maxctscut = 2
)
}
\arguments{
\item{GTsummary_list}{simply put GTSummary (Generated by CW_mgatk.read) into list, this allows mergeing multiple dataset this way.}

\item{depth_list}{simply put depth(Generated by DepthSummary) into list, this allows mergeing multiple dataset this way.}

\item{feature.list_list}{simply put feature.list(Generated by Vfilter_v3) into list, this allows mergeing multiple dataset this way.}

\item{labels}{a vector of labels for the samples.}

\item{thr}{One of the following "Total","VerySensitive","Sensitive","Specific"}

\item{qualifiedCellCut}{The minimum median mitochondrial coverage for a qualified cell, default is 10}

\item{OnlyHetero}{If only consider the heteroplasmy variants, default is T}

\item{VAFcut}{only use variants with VAF smaller than VAFcut. Default is 1.  We can use smaller value to constrain into only using rare variants}

\item{Cellcut}{only use variants with at least cellcut cells carry}

\item{maxctscut}{only use variants with at least in one cell with at leaset maxctscut variant fragments}
}
\value{
redeemR class
}
\description{
This function is to create redeemR with basic information
}
