% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{Create_redeemR}
\alias{Create_redeemR}
\title{Create_redeemR}
\usage{
Create_redeemR(
  VariantsGTSummary = VariantsGTSummary,
  qualifiedCellCut = 10,
  VAFcut = 1,
  Cellcut = 2,
  maxctscut = 2
)
}
\arguments{
\item{VariantsGTSummary}{simply put GTSummary (Generated by redeemR.read)}

\item{qualifiedCellCut}{The minimum median mitochondrial coverage for a qualified cell, default is 10}

\item{VAFcut}{only use variants with VAF smaller than VAFcut. Default is 1.  We can use smaller value to constrain into only using rare variants}

\item{Cellcut}{only use variants with at least cellcut cells carry}

\item{maxctscut}{only use variants with at least in one cell with at leaset maxctscut variant fragments}

\item{OnlyHetero}{If only consider the heteroplasmy variants, default is T}
}
\value{
redeemR class
}
\description{
This function is to create redeemR with basic information
}
