% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSummary.R
\name{CV}
\alias{CV}
\title{Internal CV}
\usage{
CV(x)
}
\arguments{
\item{x}{input a vector of numeric values}
}
\description{
This function allows you to read raw data from XX/final folder, the output from mitoV
}
