% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSummary.R
\name{CW_mgatk.read}
\alias{CW_mgatk.read}
\title{Function to read in mitoV outputs}
\usage{
CW_mgatk.read(path, Processed = F)
}
\arguments{
\item{path}{The XX/final folder, the output from mitoV}

\item{Processed}{Boolean variable (Default F), if true directly readRDS("VariantsGTSummary.RDS") or, generate and saveout "VariantsGTSummary.RDS"}
}
\value{
this returns depth which is a list of 4 df (Total/VerySensitive/Sensitive/Specific), each is a genotype summary
}
\description{
This function allows you to read raw data from XX/final folder, the output from mitoV
}
\examples{
WD<-"/lab/solexa_weissman/cweng/Projects/MitoTracing_Velocity/SecondaryAnalysis/Donor01_CD34_1_Multiomekit/MTenrichCombine/Enrich/CW_mgatk/final"
DN1CD34_1.VariantsGTSummary<-CW_mgatk.read(WD,Processed =T)
}
