% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\docType{class}
\name{Datatoplots-class}
\alias{Datatoplots-class}
\alias{Datatoplots}
\title{An intermediate  S4 class Datatoplots}
\description{
An intermediate  S4 class Datatoplots
}
\section{Slots}{

\describe{
\item{\code{clustering}}{dataframe that store the data to plot}
}}

