% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions.R
\name{DE.gettripple}
\alias{DE.gettripple}
\title{DE.gettripple}
\usage{
DE.gettripple(datapair, cpcol, withscran = F)
}
\arguments{
\item{datapair}{tyhe datapair generated from datapair.mk}

\item{cpcol}{The column name for comparison.}

\item{withscran}{if true, use deconvolution to calculate size effect.}
}
\value{
This will return .tri.dummy file that is the input for DE analysis
}
\description{
This function is to prepare the data format that is used to differentially expression calling. It include the raw matrix; data.info and size effect
}
\examples{
ROCKvsnorock.endo.tri.dummy<-DE.gettripple(ROCKvsnorock.endo.paired,cpcol="name")
}
