% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\docType{class}
\name{DistObjects-class}
\alias{DistObjects-class}
\alias{DistObjects}
\title{An intermediate  S4 class Datatoplots}
\description{
An intermediate  S4 class Datatoplots
}
\section{Slots}{

\describe{
\item{\code{jaccard}}{distance object dist: Jaccard distance}

\item{\code{Dice}}{distance object dist: Dice distance}

\item{\code{jaccard3W}}{distance object dist: jaccard3W}
}}

