% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter_redeemR_by_cells}
\alias{filter_redeemR_by_cells}
\title{Filter by cell subset}
\usage{
filter_redeemR_by_cells(
  redeemR_obj,
  cells_to_keep,
  min_cells_per_variant = 2,
  QualifiedTotalCts = NULL,
  filter_name = NULL,
  update_depth_matrix = TRUE
)
}
\arguments{
\item{redeemR_obj}{A \code{redeemR} object.}

\item{cells_to_keep}{Character vector of cell names to retain.}

\item{min_cells_per_variant}{Integer; minimum rows per variant (default 2). This is needed because after filterout some cells, some mutation may only show in 1 cell}

\item{QualifiedTotalCts}{Optional "QualifiedTotalCts" for depth rebuilding.}

\item{filter_name}{Optional label for matrix-dimension printing.}
}
\value{
The updated \code{redeemR} object.
}
\description{
Keeps only rows in \code{@GTsummary.filtered} where the cell is in \code{cells_to_keep},
enforces a minimum number of rows per variant, updates \code{@V.fitered}, and rebuilds
matrices and depth matrix.
}
