% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions.R
\name{Fun.enrich_withFC}
\alias{Fun.enrich_withFC}
\title{Fun.enrich_withFC}
\usage{
Fun.enrich_withFC(
  markergenesList,
  All.genes = all.genes.refer,
  db = msig.db,
  qcutoff = 0.05,
  top = NULL
)
}
\arguments{
\item{markergenesList}{is a list containing several elements each of which is a vector of gene names}

\item{All.genes}{the background genes  data(all.genes.refer)}

\item{db}{=msig.db   the term database to use, default is the msig.db;  data(msig.db)}

\item{qcutoff=0.05}{the term to show with q value lower than 0.05}

\item{top=NULL}{pass a number to only show top # terms}
}
\value{
this will return a simple object with two elements, one is the table to show the significant terms , the other one shows the gene names involved significant terms
}
\description{
This function is to do GSEA enrichment analysis
}
\examples{
S7G.GSEA<-Fun.enrich_withFC(markergenesList=list(S7G=topS7G.genes),All.genes=all.genes.refer)
}
