% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_functions.R
\name{GEM_Wrapper}
\alias{GEM_Wrapper}
\title{Wrap Seurat RNA clustering}
\usage{
GEM_Wrapper(mtx = bmmc.data$`Gene Expression`, exp = "DN1_BMMC1", res = 0.5)
}
\arguments{
\item{mtx}{sparse Matrix of class "dgCMatrix", each row is a gene, each column is a cell,}

\item{exp}{The name of this sample/experiment}

\item{res}{clustering resolution, default=0.5}
}
\value{
this returns seurat object with RNA clustering
}
\description{
This function allows you to perform standard sc-RNA clustering
}
\examples{
bmmc.data=Read10X(data.dir = "XX/CellRanger/Donor01_BMMC_1/outs/filtered_feature_bc_matrix")
docluster_GEM(mtx=bmmc.data$`Gene Expression`,exp="DN1_BMMC1")
}
