% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{get_ancestral_nodes}
\alias{get_ancestral_nodes}
\title{This is a convinience function, internal borrowed from https://github.com/NickWilliamsSanger/treemut/blob/main/R/treemut.R#L68}
\usage{
get_ancestral_nodes(node, edge, exclude_root = TRUE)
}
\description{
This is a convinience function, internal borrowed from https://github.com/NickWilliamsSanger/treemut/blob/main/R/treemut.R#L68
}
