% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{Get_Clonal_Variants}
\alias{Get_Clonal_Variants}
\title{Get_Clonal_Variants}
\usage{
Get_Clonal_Variants(object)
}
\arguments{
\item{object}{ReDeeM object}
}
\description{
This function identify specific mutations for each clone based on Fisher Exact Test
Of note, the ReDeeM object need to have Clone_merge in CellMeta (After running Add_tree_cut)
}
