% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_summary.R
\name{get_hype_v2}
\alias{get_hype_v2}
\title{get_hype_v2
This function annotates the redeem@V.fitered with hypermutable mutations_v2 (using 3 young donor and 2 aged donors)}
\usage{
get_hype_v2(cut = 0.01)
}
\arguments{
\item{redeemR}{a redeemR object}
}
\description{
get_hype_v2
This function annotates the redeem@V.fitered with hypermutable mutations_v2 (using 3 young donor and 2 aged donors)
}
