% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions.R
\name{Getgenenumber}
\alias{Getgenenumber}
\title{Getgenenumber}
\usage{
Getgenenumber(x, thesource)
}
\arguments{
\item{x}{x}

\item{thesource}{thesource}
}
\value{
cur.numbers
}
\description{
Internal function Getgenenumber
}
