% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_summary.R
\name{GTSummary}
\alias{GTSummary}
\title{Function to generate GTS summary}
\usage{
GTSummary(RawGenotypes, filterN = T)
}
\arguments{
\item{RawGenotypes}{Well-named "RawGenotypes.Sensitive.StrandBalance" file in function redeemR.read or CW_mgatk.read}

\item{filterN}{Boolean variable, if true filter out the variant with "N"}
}
\value{
Genotypes.summary a dataframe that summarize several metrics for each genotype
}
\description{
This function allows you to summarize the meta data for each genotyped variant
}
\examples{
Usually used inside of function CW_mgatk.read
}
