% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{LineageBiasPlot}
\alias{LineageBiasPlot}
\title{plot_npSummary to plot the lineage composition}
\usage{
LineageBiasPlot(npresult, pre)
}
\arguments{
\item{npresult}{from ProgenyMapping_np}

\item{pre}{Any short description for this plot to print with the plot}
}
\description{
plot_npSummary to plot the lineage composition
}
