% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{Make_Cells4Nodes}
\alias{Make_Cells4Nodes}
\title{Define a function to make a list, each contains the cell names for a node}
\usage{
Make_Cells4Nodes(
  tr = DN4_SLCT_HSC_w_jaccard.njtree@phylo,
  min.node.size = 10,
  max.node.fra = 0.33
)
}
\arguments{
\item{tr}{phylo object (ape)}

\item{min.node.size}{default is 10, only the nodes with more than 10 tips are included  ( # Minimum # tips in the node to be included)}

\item{max.node.fra}{default is 0.33, only consider the nodes with less than max.node.fra*total cell number (# The up limit of the node size(Fraction of all tips) to be considered)}
}
\value{
return a list each contains the cell names for a node that meets the criteria
}
\description{
Define a function to make a list, each contains the cell names for a node
}
