% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSummary.R
\name{make_ks_test_df}
\alias{make_ks_test_df}
\title{Produce KS test results, input is the redeem object
from redeemR object,}
\usage{
make_ks_test_df(redeemR)
}
\arguments{
\item{redeemR}{a redeemR object}
}
\description{
Produce KS test results, input is the redeem object
from redeemR object,
}
