% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSummary.R
\name{make_position_df_3.5}
\alias{make_position_df_3.5}
\title{make_position, but remain all the raw genotyp information
Input is the ind_df or}
\usage{
make_position_df_3.5(in_df)
}
\arguments{
\item{in_df}{raw genotype by redeemV}
}
\description{
make_position, but remain all the raw genotyp information
Input is the ind_df or
}
