% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{Make_tree,redeemR-method}
\alias{Make_tree,redeemR-method}
\title{Make_tree
This will generate a basic phylogenetic tree}
\usage{
\S4method{Make_tree}{redeemR}(object, d, algorithm, onlyreturntree = F)
}
\arguments{
\item{object}{redeemR class}

\item{d}{"jaccard" or "Dice" or "jaccard3W" or  "w_jaccard"  "w_cosine"  "LSIdist"}

\item{algorithm}{the algorithm used to build the tree, choose from "nj" and "upgma"}
}
\value{
redeemR class
}
\description{
Make_tree
This will generate a basic phylogenetic tree
}
