% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{Make_tree}
\alias{Make_tree}
\title{Make_tree
This will generate a basic phylogenetic tree}
\usage{
Make_tree(object, d = "jaccard", algorithm = "upgma", onlyreturntree = F, ...)
}
\arguments{
\item{object}{redeemR class}

\item{d}{"jaccard" or "Dice" or "jaccard3W"}

\item{algorithm}{the algorithm used to build the tree, choose from "nj" and "upgma"}
}
\description{
Make_tree
This will generate a basic phylogenetic tree
}
