% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{Make_tree_legacy,mitoTracing-method}
\alias{Make_tree_legacy,mitoTracing-method}
\title{Make_tree_legacy
This will generate a basic phylogenetic tree}
\usage{
\S4method{Make_tree_legacy}{mitoTracing}(object, d, algorithm, onlyreturntree = F)
}
\arguments{
\item{object}{redeemR class}

\item{d}{"jaccard" or "Dice" or "jaccard3W" or  "w_jaccard"  "w_cosine"  "LSIdist"}

\item{algorithm}{the algorithm used to build the tree, choose from "nj" and "upgma"}
}
\value{
redeemR class
}
\description{
Make_tree_legacy
This will generate a basic phylogenetic tree
}
