% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{Make_tree_legacy}
\alias{Make_tree_legacy}
\title{Make_tree_legacy
This will generate a basic phylogenetic tree}
\usage{
Make_tree_legacy(
  object,
  d = "jaccard",
  algorithm = "upgma",
  onlyreturntree = F,
  ...
)
}
\arguments{
\item{object}{mitoTracing class}

\item{d}{"jaccard" or "Dice" or "jaccard3W"}

\item{algorithm}{the algorithm used to build the tree, choose from "nj" and "upgma"}
}
\description{
Make_tree_legacy
This will generate a basic phylogenetic tree
}
