% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MergeMtx}
\alias{MergeMtx}
\title{Function to Merge sparse Matrix}
\usage{
MergeMtx(mtx.list, postfix)
}
\arguments{
\item{mtx.list}{A list of sparse matrix to be merged}

\item{postfix}{a vector of postfix (Usually are numbers that added at the end of cell names). Better be consistent with a merged redeemR object orders}
}
\value{
new sparse matrix
}
\description{
This function allows you to input a list of sparse matrix and merge by rownames, return a new sparse matrix
}
\examples{
Donor4_HSC_HPC_BMMC.Mtx<-MergeMtx(list(Donor04_BMMC_Multiome_wrapper$seurat@assays$RNA@counts,Donor04_HPC_Multiome_wrapper$seurat@assays$RNA@counts,Donor04_HSC_Multiome_wrapper$seurat@assays$RNA@counts),c(3,2,1))
Donor4_HSC_HPC_BMMC.RNA.seurat<-GEM_Wrapper(Donor4_HSC_HPC_BMMC.Mtx)
}
