% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\docType{class}
\name{mitoTracing-class}
\alias{mitoTracing-class}
\alias{mitoTracing}
\title{Legacy function: Major mitoTracing class that store clonal-resolved multi-omics}
\description{
Legacy function: Major mitoTracing class that store clonal-resolved multi-omics
}
\section{Slots}{

\describe{
\item{\code{GTsummary.filtered}}{The Mitochondrial genotype data frame}

\item{\code{CellMeta}}{Store meta data for each cell type}

\item{\code{V.fitered.list}}{a list of data frame of variant metrics, VAF, cellN, etc (each for different stringency),}

\item{\code{UniqueV}}{A character showing the number of usable variant}

\item{\code{Cts.Mtx}}{A sparse matrix cell-mitoVariants, store the variant count}

\item{\code{Cts.Mtx.bi}}{A sparse matrix cell-mitoVariants, The variant count has been binarized into 0 and 1}

\item{\code{Ctx.Mtx.depth}}{A sparse matrix cell-mitoVariants(total counts for each position), store the variant count}

\item{\code{para}}{A character showing the parameter of this object}

\item{\code{Seurat}}{Seurat object storing the clonal clustering results}

\item{\code{DataToplotList}}{The customized class of Datatoplots: A list of dataframe for further plotting}

\item{\code{DistObjects}}{The customized class that stores the cell-cell distances}

\item{\code{TREE}}{The customized class that wraps phylogenetic tree}
}}

