% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions.R
\name{Motifenrich.binom}
\alias{Motifenrich.binom}
\title{Motifenrich.binom
In house function to compute enrichment from Fimo
This function was developed based on HSC_multiome_Het.ipynb and HSC_multiome_Het_2.ipynb}
\usage{
Motifenrich.binom(queryP.motif, controlP.motif, alt = "greater")
}
\arguments{
\item{queryP.motif}{can be a subset of all.motif.sig}

\item{controlP.motif}{can be all.motif.sig}

\item{alt}{default is greater}
}
\description{
Motifenrich.binom
In house function to compute enrichment from Fimo
This function was developed based on HSC_multiome_Het.ipynb and HSC_multiome_Het_2.ipynb
}
