% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{Multi_Wrapper}
\alias{Multi_Wrapper}
\title{Wrap Seurat Multiomics clustering}
\usage{
Multi_Wrapper(
  path =
    "/lab/solexa_weissman/cweng/Projects/MitoTracing_Velocity/SecondaryAnalysis/Donor01_CD34_1_Multiomekit/CellRanger/Donor01_CD34_1/outs",
  atacmin = 1000,
  umimin = 1000,
  CellID = NULL
)
}
\arguments{
\item{path}{this should be the path to the cell-ranger results XX/outs}

\item{atacmin}{minimum atac fragment for each cell, default is 1000}

\item{umimin}{minimum rna umi for each cell, default is 1000}

\item{cellID}{to be used for input(useful for re-clustering), default is NULL which will use the info from path/per_barcode_metrics.csv}
}
\value{
this returns seurat object with both RNA and ATAC
}
\description{
This function allows you to perform standard sc-multiome clustering
}
\examples{
Multi_Wrapper(path="XX/CellRanger/Donor01_BMMC_1/outs/")
}
