% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{MutationProfile.bulk}
\alias{MutationProfile.bulk}
\title{Function to plot bulk level mutation signatures}
\usage{
MutationProfile.bulk(cell_variants)
}
\arguments{
\item{cell_variants}{A character vector of variants formatted as \code{position_ref_alt}, for example \code{c("93_A_G", "103_G_A")}.}
}
\value{
A ggplot object.
}
\description{
This function allows you to plot the mito mutation signatures
}
\examples{
MutationProfile.bulk(c("93_A_G", "103_G_A", "146_T_C"))
}
