% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot_depth_legacy}
\alias{plot_depth_legacy}
\title{Legacy Function to plot the mito depth summary}
\usage{
plot_depth_legacy(depth = DN1CD34_1.depth, name = "", w = 10, h = 3)
}
\arguments{
\item{depth}{The .depth file by function DepthSummary}

\item{name}{The plot name shown on top}

\item{w}{the Width of the plot, default=10}

\item{h}{the height of the plot default=3}
}
\value{
directly out put the plot
}
\description{
This function allows you to plot both position-wise and cell-wise mito depth summary
}
\examples{
plot_depth(DN1CD34_1.depth$Total,"Total")
}
