% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{plot_npSummary}
\alias{plot_npSummary}
\title{plot_npSummary to assess the outputlevel}
\usage{
plot_npSummary(npresult, orderby = "Total.norm", pre)
}
\arguments{
\item{npresult}{from ProgenyMapping_np}

\item{orderby}{Normalize by, so far can work with "Total.norm" and "Total.norm_NPadj"}

\item{pre}{Any short description for this plot to print with the plot}
}
\description{
plot_npSummary to assess the outputlevel
}
