% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_summary.R
\name{PosCoverageCellCount}
\alias{PosCoverageCellCount}
\title{Function to count cells with non-zero coverage per position}
\usage{
PosCoverageCellCount(path, CellSubset = NA)
}
\arguments{
\item{path}{The XX/final folder, the output from mitoV}

\item{CellSubset}{A vector of ATAC cell names for subsetting, default is NA}
}
\value{
A dataframe with Position and counts of cells with non-zero coverage for each threshold (Total, LS, S, VS)
}
\description{
This function creates a lookup table showing how many cells have non-zero coverage at each position for each threshold
}
\examples{
\dontrun{
WD <- "/path/to/redeemV/final"
pos_coverage <- PosCoverageCellCount(WD)
}
}
