% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_redeemR_matrix_dims}
\alias{print_redeemR_matrix_dims}
\title{Print redeemR matrix dimensions}
\usage{
print_redeemR_matrix_dims(ob, sample_name = NULL)
}
\arguments{
\item{ob}{A redeemR object}
}
\value{
Prints matrix dimensions to console
}
\description{
Displays the dimensions of all matrices in a redeemR object for quick inspection.
}
\examples{
\dontrun{
print_redeemR_matrix_dims(redeemR_object)
}

}
