% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{ProgenyMapping}
\alias{ProgenyMapping}
\title{Define a function to perform single-cell based hard porogeny assignment
This function was developed based on DN4T2.basics.ipynb}
\usage{
ProgenyMapping(
  HSC_mitoTracing = DN4_PhenoHSC_mitoTracing.verysensitive,
  Full_mitoTracing = DN4_BMMC_HSPC_HSC_mitoTracing.verysensitive,
  distCut = 0.95,
  d = "w_jaccard"
)
}
\arguments{
\item{HSC_mitoTracing}{The HSC_mitoTracing is the mitoTracing object for defined HSC}

\item{Full_mitoTracing}{The FULL_mitoTracing is the mitoTracing object for the full BMMC_HSPC_HSC}

\item{distCut}{Default is 0.95, the distance, below which I define as the related progeny}
}
\description{
Define a function to perform single-cell based hard porogeny assignment
This function was developed based on DN4T2.basics.ipynb
}
