% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{quick_w_jaccard}
\alias{quick_w_jaccard}
\title{Compute weighted jaccard distance}
\usage{
quick_w_jaccard(M, w)
}
\arguments{
\item{M}{the binary matrix, Each row is a cell, each column is a variant, generated by Make_matrix}

\item{w}{weight for each variant, a vector}
}
\value{
dist object
}
\description{
Compute weighted jaccard distance
}
