% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Reclustering}
\alias{Reclustering}
\title{Function to reclustering a seurat object}
\usage{
Reclustering(ob)
}
\arguments{
\item{ob}{a seurat object}
}
\value{
a seurat object
}
\description{
This function allows you to input a seurat object(multiome), redo clustering. Usually this is after subset
}
