% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Reclustering_hm}
\alias{Reclustering_hm}
\title{Function to reclustering_hm a seurat object with Harmony}
\usage{
Reclustering_hm(
  ob = DN4_RigHSC_T1T2_Multiome_wrapper_filtered.anno,
  HarmonyBy = "TimePoint"
)
}
\arguments{
\item{ob}{a seurat object}

\item{HarmonyBy}{The columne name in meta that will be used for Harmony}
}
\value{
a seurat object
}
\description{
This function allows you to input a seurat object(multiome), redo clustering harmony by a certain column in meta data. Usually this is after subset
}
