% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{reconstruct_genotype_summary}
\alias{reconstruct_genotype_summary}
\title{This is a function borrowed from https://github.com/NickWilliamsSanger/treemut/blob/main/R/treemut.R#L68
Input phylo object, return a "profile matrix"--Edge(or denoted as the ending node) vs cell. a 0, 1 character string that indicate what cells in a given node}
\usage{
reconstruct_genotype_summary(phylo)
}
\arguments{
\item{phylo}{phylo an ape object}
}
\value{
df  includes df$df which is a big data frame,  and df$sample that is the cell names
}
\description{
This is a function borrowed from https://github.com/NickWilliamsSanger/treemut/blob/main/R/treemut.R#L68
Input phylo object, return a "profile matrix"--Edge(or denoted as the ending node) vs cell. a 0, 1 character string that indicate what cells in a given node
}
