% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_summary.R
\name{redeemR.read.multiple.trim}
\alias{redeemR.read.multiple.trim}
\title{Function to read in multiple runs of redeemV outputs with edge trimming, default is trimming 5bp}
\usage{
redeemR.read.multiple.trim(paths, names, suffix, edge_trim = 5)
}
\arguments{
\item{paths}{a vector of multiple path strings The XX/final folder, the output from mitoV}

\item{names}{The names for the multiple input, same order with folder}

\item{suffix}{The suffix for the multiple input, which is added at the end of cell name, separated by "."}

\item{edge_trim}{how many bp to be trimmed, default is 5,}
}
\value{
VariantsGTSummary  combined VariantsGTSummary
}
\description{
This function allows you to read raw data from multiple XX/final folder, the output from redeemV
It process the data same way as CW_mgatk.read but need to specify one threadhold(thr)
}
