% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{run_redeem_qc}
\alias{run_redeem_qc}
\title{run_redeem_qc}
\usage{
run_redeem_qc(redeem, homosets, hotcall = c("310_T_C", "9979_G_A", "3109_T_C"))
}
\arguments{
\item{redeem}{The redeemR object}

\item{homosets}{the homoplasmic mutation set}

\item{hotcall}{make sure hotcall mutations are not included}
}
\description{
This function generate qc plot to assess filtering strategy
}
