% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{Runplot_scale_2}
\alias{Runplot_scale_2}
\title{plot_npSummary to assess the outputlevel vs lineage bias, normalize by assigned}
\usage{
Runplot_scale_2(
  datatoplot = DN4_HSC_LSI_progeny$LineageSummary$output_lineage.summary.pct.scale,
  pre
)
}
\arguments{
\item{datatoplot}{A slot from the result of ProgenyMapping_np : datatoplot.scale}

\item{pre}{Any short description for this plot to print with the plot}
}
\description{
plot_npSummary to assess the outputlevel vs lineage bias, normalize by assigned
}
