% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_dim}
\alias{safe_dim}
\title{Safely get matrix-like dimensions}
\usage{
safe_dim(x)
}
\arguments{
\item{x}{A matrix/data.frame/Matrix object or NULL.}
}
\value{
Integer vector of length 2: c(nrow, ncol), or c(NA_integer_, NA_integer_) if x is NULL.
}
\description{
Returns a pair of integers for the number of rows and columns of a
matrix-like object. If the input is NULL, returns NA for both values.
}
