% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{SeuratLSIClustering,mitoTracing-method}
\alias{SeuratLSIClustering,mitoTracing-method}
\title{SeuratLSIClustering
This will use the mito variants for Seurat clustering (LSI based)}
\usage{
\S4method{SeuratLSIClustering}{mitoTracing}(object, binary = T, res = 0.3)
}
\arguments{
\item{binary}{Default is tree, to make use of the binary matrix}

\item{res}{Default os 0.3, the resolution of the clustering}

\item{mitoTracing}{class}
}
\value{
mitoTracing class
}
\description{
SeuratLSIClustering
This will use the mito variants for Seurat clustering (LSI based)
}
