% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{SeuratLSIClustering,redeemR-method}
\alias{SeuratLSIClustering,redeemR-method}
\title{SeuratLSIClustering
This will use the mito variants for Seurat clustering (LSI based)}
\usage{
\S4method{SeuratLSIClustering}{redeemR}(
  object,
  binary = T,
  res = 0.6,
  lsidim = 2:50,
  rmvariants = c("Variants310TC", "Variants3109TC", "Variants5764CT")
)
}
\arguments{
\item{binary}{Default is tree, to make use of the binary matrix}

\item{res}{Default os 0.3, the resolution of the clustering}

\item{redeemR}{class}
}
\value{
redeemR class
}
\description{
SeuratLSIClustering
This will use the mito variants for Seurat clustering (LSI based)
}
