% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{show,mitoTracing-method}
\alias{show,mitoTracing-method}
\title{show
This will show the basics of mitoTracin class}
\usage{
\S4method{show}{mitoTracing}(object)
}
\arguments{
\item{object}{mitoTracin class}
}
\value{
print out basics
}
\description{
show
This will show the basics of mitoTracin class
}
