% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{show,redeemR-method}
\alias{show,redeemR-method}
\title{show
This will show the basics of redeemR class}
\usage{
\S4method{show}{redeemR}(object)
}
\arguments{
\item{object}{redeemR class}
}
\value{
print out basics
}
\description{
show
This will show the basics of redeemR class
}
