% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Show_Consensus}
\alias{Show_Consensus}
\title{Function to plot consensus mtDNA mutation benchmark}
\usage{
Show_Consensus(ob, N = 25)
}
\arguments{
\item{ob}{The redeemR object}

\item{N}{number of example variants to show, default is 25}
}
\description{
This function allows you to plot the mito mutation consensus levels
It will print out Quantiles of UMI family size; Quantile of consensus score; Percentage of R1/R2 overlaped mutation detections
It will also plot random N mutations as examples to show consensus metrics
}
