% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{split_profile}
\alias{split_profile}
\title{This is a convinience function, internal  borrowed from https://github.com/NickWilliamsSanger/treemut/blob/main/R/treemut.R#L68}
\usage{
split_profile(profile)
}
\description{
This is a convinience function, internal  borrowed from https://github.com/NickWilliamsSanger/treemut/blob/main/R/treemut.R#L68
}
