% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{Subset_MitoTracing}
\alias{Subset_MitoTracing}
\title{Subset_MitoTracing Subset a mitotracing object by selecting a subset of cells, return a new MitoTracing object with only 4 slots: para; CellMeta; Cts.Mtx.bi; UniqueV, can be used for downstreme compute distance, clonal clustering, make tree, etc}
\usage{
Subset_MitoTracing(MitoTracing, Cells, ExtraInfo = "Subset from ... ")
}
\arguments{
\item{Cells}{Important, give a vector of Cell names(ATAC cell names)}

\item{ExtraInfo}{Extra information, usually "Subset from ..."}

\item{Mitotracing}{The Parent MitoTracing object eg. DN4_HSC_mitoTracing.Sensitive}
}
\value{
MitoTracing Object
}
\description{
Subset_MitoTracing Subset a mitotracing object by selecting a subset of cells, return a new MitoTracing object with only 4 slots: para; CellMeta; Cts.Mtx.bi; UniqueV, can be used for downstreme compute distance, clonal clustering, make tree, etc
}
