% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_cw1}
\alias{theme_cw1}
\title{Define a custom theme function}
\usage{
theme_cw1(axis_title_size = 20, axis_text_size = 15)
}
\arguments{
\item{axis_title_size}{}

\item{axis_text_size}{}
}
\description{
Define a custom theme function
}
