% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Translate_simple_RNA2ATAC}
\alias{Translate_simple_RNA2ATAC}
\title{Translate_simple_RNA2ATAC}
\usage{
Translate_simple_RNA2ATAC(
  name,
  PostFix = T,
  bclength = 16,
  from = c(1, 2, 3),
  to = c(1, 2, 3)
)
}
\arguments{
\item{name}{RNA name, as the RNA cell barcode usually with the post -1}

\item{bclength}{The cell barcode length, default is 16}

\item{from}{A vector of the postfix,  usually is c(1,2,3,...), it depends on how many samples are aggregated in Cellranger RNA part}

\item{to}{A vector of the postfix, those cooresponds to the postfix added in redeemR, in general, if it matches, then simply c(1,2,3,...),
but in case not match, here provides a way to transform into redeemR order}
}
\value{
ATAC name
Translate_RNA2ATAC(\verb{a vector of RNA names})
}
\description{
This function allows you to input the RNA name to translate to ATAC name
}
