% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\docType{class}
\name{TREE-class}
\alias{TREE-class}
\alias{TREE}
\title{An intermediate  S4 class Tree that store tree info}
\description{
An intermediate  S4 class Tree that store tree info
}
\section{Slots}{

\describe{
\item{\code{phylo}}{the phylo tree class from ape package}

\item{\code{treedata}}{treedata class from tidytree}

\item{\code{records}}{character to store annotations}
}}

