% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSummary.R
\name{Vfilter_v4}
\alias{Vfilter_v4}
\title{Function to filter variants, v4}
\usage{
Vfilter_v4(
  InputSummary = VariantsGTSummary,
  Min_Cells = 2,
  Max_Count_perCell = 2,
  QualifyCellCut = 10
)
}
\arguments{
\item{InputSummary}{The GTSummary file read in by function CW_mgatk.read}

\item{Min_Cells}{Default 2, A qualified variant needs the minimum number of cells that have this variant}

\item{Max_Count_perCell}{Default 2,  A qualified variant needs to show at least 2 counts in one cell}

\item{QualifyCellCut}{Default 10, Minimum depth for a qualified cell}

\item{depth}{The .depth file by function DepthSummary}

\item{Rmvhomo}{Boolean (Default F) If true, remove the homozygous variants}
}
\value{
this returns feature.list
}
\description{
This function allows you to filter variants,deprecated, use Vfilter_v4 instead
}
\examples{
DN1CD34_1.Variants.feature.lst<-Vfilter_v3(InputSummary=DN1CD34_1.VariantsGTSummary,depth=DN1CD34_1.depth)
}
