% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions.R
\name{add_changes}
\alias{add_changes}
\title{add_changes}
\usage{
add_changes(variant)
}
\arguments{
\item{variant}{given a variant, output the changes}
}
\value{
changes
}
\description{
add_changes
}
