% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{Add_DepthMatrix_filter2}
\alias{Add_DepthMatrix_filter2}
\title{Add Filter-2–adjusted depth matrix to a redeemR object}
\usage{
Add_DepthMatrix_filter2(object, QualifiedTotalCts = NULL)
}
\description{
Add Filter-2–adjusted depth matrix to a redeemR object
}
