% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_summary.R
\name{add_freq_raw}
\alias{add_freq_raw}
\title{Convinient function that takes raw fragment, and out put fragment with frequency
Produce KS test results, input is the redeem object
from redeemR object,}
\usage{
add_freq_raw(raw)
}
\description{
Convinient function that takes raw fragment, and out put fragment with frequency
Produce KS test results, input is the redeem object
from redeemR object,
}
