% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{Add_tree_cut}
\alias{Add_tree_cut}
\title{Add_tree_cut
a function to cut tree using assigned variant as branch-length on edge}
\usage{
Add_tree_cut(redeemR, MinCell, N, ...)
}
\arguments{
\item{redeemR}{Need to have had the tree built}

\item{MinCell}{The minimum number of cells in each clone, otherwise merge with sibling}

\item{N}{branch length to cut the tree}
}
\description{
Add_tree_cut
a function to cut tree using assigned variant as branch-length on edge
}
