% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions.R
\name{add_types}
\alias{add_types}
\title{add_types}
\usage{
add_types(changes)
}
\arguments{
\item{changes}{given a changes, output the type (transition or transversion)}
}
\value{
changes
}
\description{
add_types
}
