% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{AddDatatoplot_clustering,redeemR-method}
\alias{AddDatatoplot_clustering,redeemR-method}
\title{AddDatatoplot_clustering
This prepare the clonal clustering data to plot}
\usage{
\S4method{AddDatatoplot_clustering}{redeemR}(object)
}
\arguments{
\item{object}{redeemR class}
}
\value{
redeemR class
}
\description{
AddDatatoplot_clustering
This prepare the clonal clustering data to plot
}
