% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_tree.R
\name{AddDist}
\alias{AddDist}
\title{AddDist
This add Jaccard, Dice, Jaccard3W distance and stored in DistObjects}
\usage{
AddDist(object, ...)
}
\arguments{
\item{object}{redeemR class}
}
\description{
AddDist
This add Jaccard, Dice, Jaccard3W distance and stored in DistObjects
}
