% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions.R
\name{Annotate_base_change}
\alias{Annotate_base_change}
\title{Annotate_base_change}
\usage{
Annotate_base_change(redeem)
}
\arguments{
\item{input}{redeem object,  it takes V.fitered, add the nucleotide change,}
}
\value{
redeem object with the V.fitered slot modified
}
\description{
Annotate_base_change
}
