% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{BinaryDist}
\alias{BinaryDist}
\title{Compute distances for binary distances}
\usage{
BinaryDist(M, method = "Jaccard")
}
\arguments{
\item{M}{the binary matrix, Each row is a cell, each column is a variant, generated by Make_matrix}

\item{method}{distance method, choose from Jaccard, Dice, 3WJaccard, Simpson, Kulczynski2, Ochiai, Hamming}
}
\value{
dist object
}
\description{
Compute distances for binary distances
}
\examples{
d.Jaccard<-BinaryDist(object@Cts.Mtx.bi,method="Jaccard")
}
