% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSummary.R
\name{ComputeRejectRate}
\alias{ComputeRejectRate}
\title{Function to compute the reject rate(The filtering rate in concensus variant calling)}
\usage{
ComputeRejectRate(ob)
}
\arguments{
\item{ob}{The redeemR object}
}
\value{
a modified ob with RejectRate added to @CellMeta
}
\description{
This function allows you to computae the filtering rate for each single cell
}
