% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_format.R
\name{convert_redeem_matrix_long}
\alias{convert_redeem_matrix_long}
\title{Included in redeemR-2.0  2025-07-01}
\usage{
convert_redeem_matrix_long(
  sample = "",
  mat_var,
  mat_depth,
  cell_whitelist = NULL
)
}
\description{
Included in redeemR-2.0  2025-07-01
}
