% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{CountVperCell}
\alias{CountVperCell}
\title{Internal function in plot_variant}
\usage{
CountVperCell(x, name, CellN)
}
\arguments{
\item{x}{CellVar.Sum$VN}

\item{name}{c}

\item{CellN}{nrow(CellVar.Sum)}
}
\description{
Internal function in plot_variant
}
\examples{
CountVperCell(CellVar.Sum$VN,c,CellN=nrow(CellVar.Sum))
}
