% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{Create_redeemR_model}
\alias{Create_redeemR_model}
\title{Create_redeemR_model}
\usage{
Create_redeemR_model(
  VariantsGTSummary = VariantsGTSummary,
  qualifiedCellCut = 10,
  VAFcut = 1,
  Cellcut = 2
)
}
\arguments{
\item{VariantsGTSummary}{simply put GTSummary (Generated by redeemR.read)}

\item{qualifiedCellCut}{The minimum median mitochondrial coverage for a qualified cell, default is 10}

\item{VAFcut}{only use variants with VAF smaller than VAFcut. Default is 1.  We can use smaller value to constrain into only using rare variants}

\item{Cellcut}{only use variants with at least cellcut cells carry}

\item{OnlyHetero}{If only consider the heteroplasmy variants, default is T}
}
\value{
redeemR class
}
\description{
This function is to create redeemR with basic information, modeled by binomial distribution
maxctscut is removed, replaced by the modeling, will generate statistics and p value
}
