% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{Make_matrix,redeemR-method}
\alias{Make_matrix,redeemR-method}
\title{Make_matrix
This will make the matixies of Cell VS mitochondrial variants and return redeemR
Results stored in Cts.Mtx and Cts.Mtx.bi}
\usage{
\S4method{Make_matrix}{redeemR}(object, onlyhetero = T)
}
\arguments{
\item{object}{redeemR class}

\item{onlyhetero}{Only use heteroplasmic mutations}
}
\value{
redeemR class
}
\description{
Make_matrix
This will make the matixies of Cell VS mitochondrial variants and return redeemR
Results stored in Cts.Mtx and Cts.Mtx.bi
}
