% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{Make_matrix}
\alias{Make_matrix}
\title{Make_matrix
This will make the matixies of Cell VS mitochondrial variants and return mitoTracing
Results stored in Cts.Mtx and Cts.Mtx.bi}
\usage{
Make_matrix(object)
}
\arguments{
\item{object}{mitoTracin class}
}
\description{
Make_matrix
This will make the matixies of Cell VS mitochondrial variants and return mitoTracing
Results stored in Cts.Mtx and Cts.Mtx.bi
}
