% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{MakeAllNodes}
\alias{MakeAllNodes}
\title{Define a function make the Allnodes(Node|Parent|Freq|CladeSize), where Freq is the number of variants assigned to the node(as ending point)
from redeemR object,}
\usage{
MakeAllNodes(redeemR = DN4_stemcell_redeemR.seed.verysensitive, prob.cut = 0.3)
}
\arguments{
\item{redeemR}{a redeemR object already have the tree built}

\item{prob.cut}{The probability cutoff to include confidently assigned variant}
}
\description{
Define a function make the Allnodes(Node|Parent|Freq|CladeSize), where Freq is the number of variants assigned to the node(as ending point)
from redeemR object,
}
