% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{MakeDF4Regress}
\alias{MakeDF4Regress}
\title{MakeDF4Regress
Define a function to make two dataframe for regression analysis
This function was developed based on HSC_multiome_Het_2.ipynb}
\usage{
MakeDF4Regress(
  multiome_wrapper = Donor04_HSC_Multiome_wrapper,
  mitoTracing = DN4_stemcell_mitoTracing.seed.sensitive,
  progeny_np = DN4_HSC_LSI_progeny_np,
  assay = "SCT",
  useNPimputation = T,
  maxcloneUMI = 10
)
}
\arguments{
\item{multiome_wrapper}{This outject should includes all and more than HSCs cells in mitoTracing}

\item{mitoTracing}{scMitoTracing object for HSC}

\item{progeny_np}{run via ProgenyMapping_np}

\item{assay}{SCT for expression, ATAC for ATAC}

\item{useNPimputation}{default is T, use all cells called by network propagation, inaddition to the top cells in mitoTracing}

\item{maxcloneUMI}{default is 10, Only include genes, in the max clone the expression greater than 10}
}
\value{
list(mtx.clone=mtx.clone,mtx.clone.norm.scale=mtx.clone.norm.scale)
}
\description{
MakeDF4Regress
Define a function to make two dataframe for regression analysis
This function was developed based on HSC_multiome_Het_2.ipynb
}
