% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{MutationProfile.bulk}
\alias{MutationProfile.bulk}
\title{Function to plot bulk level mutation signatures}
\usage{
MutationProfile.bulk(cell_variants)
}
\arguments{
\item{cell_variants}{a vector of variants formated as c('93_A_G''103_G_A''146_T_C'}
}
\value{
p from ggplot2
}
\description{
This function allows you to plot the mito mutation signatures
}
\examples{
MutationProfile.bulk(DN1CD34_1.Variants.feature.lst[[name]]$Variants
}
