% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot_variant}
\alias{plot_variant}
\title{Function to plot variant metrics}
\usage{
plot_variant(ob, p4xlim = 50, QualifyCellCut = 10)
}
\arguments{
\item{ob}{The redeemR object}

\item{p4xlim}{the p4 xlim(number of variant per cell), default is 50}

\item{QualifyCellCut}{median coverage for qualified cells, default is 10}
}
\value{
no returns, directly plot
}
\description{
This function allows you to plot the mito mutation metrics
For each category(stringency),
p1: Variant allele frequency(VAF);
p2: Heteroplasmy histogram
p3: CellN(Number of caells that carry the variants) VS maxcts( The number of variant counts in the highest cell)
p4: Histogram to show the distribution of the number of variant per cell
}
