% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuidTree.R
\name{SeuratLSIClustering}
\alias{SeuratLSIClustering}
\title{SeuratLSIClustering
This will use the mito variants for Seurat clustering (LSI based)}
\usage{
SeuratLSIClustering(object, ...)
}
\arguments{
\item{object}{redeemR class}
}
\description{
SeuratLSIClustering
This will use the mito variants for Seurat clustering (LSI based)
}
