% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Tomerge_v2}
\alias{Tomerge_v2}
\title{Tomerge_v2}
\usage{
Tomerge_v2(A, B, leavex = T, leavey = F)
}
\arguments{
\item{A}{dataframe A}

\item{B}{dataframe B}
}
\value{
return a data frame with merged information
}
\description{
This function is to quickly merge two dataframe by rownames, but can choose to leave A or B all information
}
\examples{
Tomerge_v2(A,B)
}
